fix(application): cast Bsize to int64 so linux/arm builds

+++ internal/application/diskfree_linux.go
@@ -11,5 +11,5 @@ func freeDiskBytes(path string) (int64, error) {
 	if err := syscall.Statfs(path, &stat); err != nil {
 		return 0, err
 	}
-	return int64(stat.Bavail) * stat.Bsize, nil
+	return int64(stat.Bavail) * int64(stat.Bsize), nil
 }
